requirements.txt
Flask==2.0.1
pycryptodome==3.10.1
PyMySQL==1.0.2
pytest==6.2.5
Flask-WTF==1.0.0
email-validator==1.1.3
pytest-cov==3.0.0